Move Oracle related operations to its own module
Anything oracle related in @chainlink/eth-test-helpers has been moved
to its own module at oracle.ts
