+++ eth-test-helpers/src/contracts/index.ts
@@ -1,0 +1,2 @@
+import * as oracle from './oracle'
+export { oracle }
