+++ eth-test-helpers/src/index.ts
@@ -6,4 +6,5 @@ import * as matchers from './matchers'
 import * as setup from './setup'
 import * as wallet from './wallet'
 
+export * from './contracts'
 export { contract, helpers, debug, wallet, matchers, setup, extensions }
