+++ evm/test/BasicConsumer.test.ts
@@ -2,6 +2,7 @@ import {
   contract,
   helpers as h,
   matchers,
+  oracle,
   setup,
 } from '@chainlink/eth-test-helpers'
 import cbor from 'cbor'
@@ -70,7 +71,7 @@ describe('BasicConsumer', () => {
         const log = receipt?.logs?.[3]
         assert.equal(log?.address ?? '', oc.address)
 
-        const request = h.decodeRunRequest(log)
+        const request = oracle.decodeRunRequest(log)
         const expected = {
           path: ['USD'],
           get:
@@ -94,17 +95,17 @@ describe('BasicConsumer', () => {
 
   describe('#fulfillOracleRequest', () => {
     const response = ethers.utils.formatBytes32String('1,000,000.00')
-    let request: h.RunRequest
+    let request: oracle.RunRequest
 
     beforeEach(async () => {
       await link.transfer(cc.address, h.toWei('1'))
       const tx = await cc.requestEthereumPrice(currency)
       const receipt = await tx.wait()
-      request = h.decodeRunRequest(receipt.logs?.[3])
+      request = oracle.decodeRunRequest(receipt.logs?.[3])
     })
 
     it('records the data given to it by the oracle', async () => {
-      await h.fulfillOracleRequest(
+      await oracle.fulfillOracleRequest(
         oc.connect(roles.oracleNode),
         request,
         response,
@@ -115,7 +116,7 @@ describe('BasicConsumer', () => {
     })
 
     it('logs the data given to it by the oracle', async () => {
-      const tx = await h.fulfillOracleRequest(
+      const tx = await oracle.fulfillOracleRequest(
         oc.connect(roles.oracleNode),
         request,
         response,
@@ -128,23 +129,23 @@ describe('BasicConsumer', () => {
     })
 
     describe('when the consumer does not recognize the request ID', () => {
-      let otherRequest: h.RunRequest
+      let otherRequest: oracle.RunRequest
 
       beforeEach(async () => {
-        const args = h.requestDataBytes(
+        const args = oracle.encodeOracleRequest(
           specId,
           cc.address,
           basicConsumerFactory.interface.functions.fulfill.sighash,
           43,
           '0x0',
         )
-        const tx = await h.requestDataFrom(oc, link, 0, args)
+        const tx = await link.transferAndCall(oc.address, 0, args)
         const receipt = await tx.wait()
-        otherRequest = h.decodeRunRequest(receipt.logs?.[2])
+        otherRequest = oracle.decodeRunRequest(receipt.logs?.[2])
       })
 
       it('does not accept the data provided', async () => {
-        await h.fulfillOracleRequest(
+        await oracle.fulfillOracleRequest(
           oc.connect(roles.oracleNode),
           otherRequest,
           response,
@@ -170,13 +171,13 @@ describe('BasicConsumer', () => {
 
   describe('#cancelRequest', () => {
     const depositAmount = h.toWei('1')
-    let request: h.RunRequest
+    let request: oracle.RunRequest
 
     beforeEach(async () => {
       await link.transfer(cc.address, depositAmount)
       const tx = await cc.requestEthereumPrice(currency)
       const receipt = await tx.wait()
-      request = h.decodeRunRequest(receipt.logs?.[3])
+      request = oracle.decodeRunRequest(receipt.logs?.[3])
     })
 
     describe('before 5 minutes', () => {
